refactor(tui): move demo implementation out of cli
