--- PRODUCT.md
@@ -1,51 +1,0 @@
-# Product
-
-## Register
-
-product
-
-## Users
-
-EverOS serves developers, agent builders, and AI product teams who need a memory
-layer they can install, inspect, and trust from day one. They usually arrive at
-the repository from PyPI, GitHub, or launch content, then decide quickly whether
-the library feels real enough to try.
-
-## Product Purpose
-
-EverOS is a Python library and local-first memory runtime for AI agents and user
-chats. It turns conversations, files, and agent trajectories into readable
-Markdown, then keeps SQLite and LanceDB indexes in sync for retrieval. The demo
-exists as a low-friction activation hook: users should be able to run one command
-and feel the memory lifecycle before wiring API keys or a server.
-
-## Brand Personality
-
-Precise, intelligent, and tactile. EverOS should feel like a serious memory
-system with a distinctive terminal-native identity, not a novelty CLI.
-
-## Anti-references
-
-Avoid generic SaaS dashboards, noisy hacker-terminal tropes, cheap ASCII art,
-decorative motion without state meaning, and demo flows that require users to
-understand the full server stack before they see value.
-
-## Design Principles
-
-Show the memory lifecycle, do not only describe it.
-
-Make the first successful moment immediate, source-backed, and visible.
-
-Treat terminal UI as a product surface with hierarchy, rhythm, and restraint.
-
-Keep the poster palette consistent: dark graphite surfaces, warm ink, and bright
-EverOS yellow as the main activation signal.
-
-Separate educational demo behavior from production server behavior so users know
-what is simulated and what is API-backed.
-
-## Accessibility & Inclusion
-
-Respect terminal contrast and readable copy. Motion should communicate state and
-have a static fallback through `everos demo --plain` for non-interactive or
-limited terminals.
