+++ src/everos/entrypoints/tui/demo/__init__.py
@@ -1,0 +1,1 @@
+"""EverOS educational demo TUI."""
