+++ src/everos/entrypoints/tui/demo/data.py
