+++ src/everos/entrypoints/tui/demo/widgets/__init__.py
@@ -1,0 +1,1 @@
+"""Widgets and rendering primitives for the EverOS demo TUI."""
