+++ src/everos/entrypoints/tui/demo/widgets/sphere.py
