+++ tests/unit/test_entrypoints/test_tui/__init__.py
@@ -1,0 +1,1 @@
+"""TUI entrypoint tests."""
